Added a helper file to make app.jsx look clean
